backend/requirements.txt
psycopg[binary]>=3.3.3,<3.4
PyJWT[crypto]>=2.12.1,<3.0
python-dotenv>=1.2.2,<2.0
